refactor(extract): 收口WAV sidecar总结与模块边界
Why:
- 需要把 WAV sidecar 的 breaker 降级说明与报告结果真正纳入 extraction 收口
- 需要补齐最终回归覆盖，并把 wav_sidecar 主文件压回项目允许的体量

What:
- 在 unpack 收口中追加 WAV sidecar summary、breaker note 与最终日志等级判定
- 为 run_summary 与 sidecar 报告写出补充回归测试，并修正默认值断言与 CLI 兼容兜底
- 新增 wav_sidecar_runtime 模块，拆出低层数据结构与 attempt runner，使 wav_sidecar.py 降到 291 行

Risk:
- 兼容性影响：低，未启用 WAV 时保持原有 extraction 结果与 CLI 行为
- 潜在风险：summary 现在会在启用 WAV 且存在失败/跳过时升级为 warning，需要后续真实样本再观察文案是否足够清晰

Test:
- uv run pytest -q tests/test_app_context.py tests/test_cli.py tests/test_wav_sidecar.py tests/test_unpack_wav_sidecar.py tests/test_run_summary.py
- uv run ruff check src/lol_audio_unpack/app_context.py src/lol_audio_unpack/__main__.py src/lol_audio_unpack/facade.py src/lol_audio_unpack/unpack.py src/lol_audio_unpack/wav_sidecar.py src/lol_audio_unpack/wav_sidecar_runtime.py tests/test_app_context.py tests/test_cli.py tests/test_wav_sidecar.py tests/test_unpack_wav_sidecar.py tests/test_run_summary.py
