+++ tests/test_app_context.py
@@ -20,6 +20,9 @@ from lol_audio_unpack.utils.runtime_paths import detect_runtime_paths
 pytestmark = pytest.mark.unit
 
 DEFAULT_MAX_WORKERS = 4
+DEFAULT_WAV_WORKER_COUNT = 2
+DEFAULT_WAV_TIMEOUT_SECONDS = 5
+DEFAULT_WAV_MAX_RETRIES = 3
 
 
 def _write_env_file(env_dir: Path, game_path: Path, output_path: Path) -> None:
@@ -668,6 +671,6 @@ def test_operation_options_defaults() -> None:
     assert options.process_events is True
     assert options.integrate_data is False
     assert options.wav_output.enabled is False
-    assert options.wav_output.worker_count == 2
-    assert options.wav_output.timeout_seconds == 5
-    assert options.wav_output.max_retries == 3
+    assert options.wav_output.worker_count == DEFAULT_WAV_WORKER_COUNT
+    assert options.wav_output.timeout_seconds == DEFAULT_WAV_TIMEOUT_SECONDS
+    assert options.wav_output.max_retries == DEFAULT_WAV_MAX_RETRIES
